Updated I2C Slave, BOD, UART and SSP drivers.
Updated I2C Slave driver. Register read does not invoke user callback.
Fixed interrupt handling in PinInt gen2 driver.
Added KConfig options for BOD and WDT disabling is deep sleep/power down.
Renamed Makefile.

Fixed SSP and UART clock when AHB div is not 1, applies to LPC11, LPC11E,
LPC13, LPC13U series. SSP and UART prescaler is set equal to AHB prescaler.

+++ include/halm/platform/lpc/i2c_slave.h
@@ -14,7 +14,7 @@ extern const struct InterfaceClass * const I2CSlave;
 
 struct I2CSlaveConfig
 {
-  /** Mandatory: register map size. Map size is limited by 64 kB. */
+  /** Mandatory: register map size. Map size is limited to 256 bytes. */
   size_t size;
   /** Mandatory: serial clock pin. */
   PinNumber scl;
@@ -36,11 +36,11 @@ struct I2CSlave
   /* Register cache */
   uint8_t *cache;
   /* Cache size */
-  uint16_t size;
+  uint8_t size;
   /* Position in register cache for external access */
-  uint16_t external;
+  uint8_t external;
   /* Position in register cache for program access */
-  uint16_t internal;
+  uint8_t internal;
 
   /* Current interface state */
   uint8_t state;
